refactor: organize strutucture and review readme
- Implemented `getProductsFromState` to extract products from various page types.
- Created `forEachProduct` higher-order function for rendering components based on products.
- Added state selectors: `getCart`, `getCartItems`, `getPageType`, and `getCustomer` for easier state access.
- Developed UI helper functions for showing toasts, clearing slots, and rendering components.
- Introduced utility functions for deep cloning, debouncing, and throttling.
- Added tests for selectors, UI helpers, and utility functions to ensure functionality.
- Removed outdated type guard tests and implementations.
