+++ packages/helper/src/lib/guards/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @fileoverview Public type guards for NubeSDK.
+ *
+ * Note: `primitives.ts` (isObject/isPlainObject) is intentionally not
+ * re-exported — those are internal building blocks, not public API.
+ */
+
+export * from "./address";
+export * from "./cart";
+export * from "./components";
+export * from "./domain";
+export * from "./pages";
